:sparkles: Support Multiple Output Formats for Decoded Tokens
