+++ jwt_debugger/command.py
@@ -5,6 +5,7 @@ from functools import partial
 
 from rich import print  # pylint: disable=redefined-builtin
 from click import File
+from click import Choice
 from click import option
 from click import command
 from click import argument
@@ -13,6 +14,8 @@ from click.core import Context
 from click.core import Argument
 from click.exceptions import UsageError
 
+from jwt_debugger.console import JSONDecodedToken
+from jwt_debugger.console import PrettyDecodedToken
 from jwt_debugger.decoder import decode_token
 from jwt_debugger.decoder import load_jwk_from_file
 from jwt_debugger.decoder import load_jwkset_from_oidc_url
@@ -26,11 +29,11 @@ def read_token_argument(unused_context :Context, unused_argument :Argument, valu
     return value
 
 
-@command()
-@option('--public-key', type=File(), help='JSON Web Key in JSON or PEM format for signature verification')
-@option('--oidc-provider-url', help='OpenID Connect Provider URL where JSON Web Key Set can be pulled for signature verification')
+@option('--public-key', type=File(), help='JSON Web Key in JSON or PEM format for signature verification.')
+@option('--oidc-provider-url', help='OpenID Connect Provider URL where JSON Web Key Set can be pulled for signature verification.')
+@option('--format', type=Choice(['pretty', 'json']), default='pretty', help='Output format')
 @argument('token', required=False, callback=read_token_argument)
-def cli(token :str, public_key :Optional[TextIOWrapper] =None, oidc_provider_url :str =None) -> None:
+def cli(token :str, format :str, public_key :Optional[TextIOWrapper] =None, oidc_provider_url :str =None) -> None:
     if all([public_key, oidc_provider_url]):
         raise UsageError('The following options can not be used together (--public-key, --oidc-provider-url).')
 
@@ -52,8 +55,13 @@ def cli(token :str, public_key :Optional[TextIOWrapper] =None, oidc_provider_url
         decode_token_ = partial(decode_token, public_key=None)
 
     decoded_token = decode_token_(token)
+    if format == 'json':
+        console_renderable = JSONDecodedToken(decoded_token.header, decoded_token.payload)
 
-    print(decoded_token)
+    else:
+        console_renderable = PrettyDecodedToken(decoded_token.token, decoded_token.header, decoded_token.payload, decoded_token.verified)
+
+    print(console_renderable)
 
     if decoded_token.verified is False:
         sys.exit(1)
